extract utils into seperate file
